requirements.txt
pillow
psutil
python-dotenv